+++ indexer/drawing_rule_def.hpp
@@ -34,7 +34,7 @@ namespace drule
   /// geo type of rule - can be one combined of ...
   enum rule_geo_t { node = 1, way = 2 };
 
-  int const layer_base_priority = 2000;
+  double const layer_base_priority = 2000;
 
   void MakeUnique(vector<Key> & keys);
   void SortByScaleTypeDepth(vector<Key> & keys);
